Added the docstring documentation
