+++ laser/README.md
@@ -1,7 +1,7 @@
 # IARC 2019 Laser Class
 
 ## Usage:
- Make a laser object and give it the pin number it is plugged into. Then you can obj.on() and obj.off() to turn it on and off (respectively).
+ Make a laser object and give it the pin number it is plugged into. Then you can 'obj.on()' and 'obj.off()' to turn it on and off (respectively).
 
 ### Notes
- Since we are using RPi.GPIO with the UpBoard you need to use this specific [kernal](https://wiki.up-community.org/RPi.GPIO).
+ Since we are using RPi.GPIO with the UpBoard you need to use this specific [kernel](https://wiki.up-community.org/RPi.GPIO).
